rTrorrent with c-ares

+++ src/scripts/boot.sh
@@ -1,26 +1,25 @@
 #!/bin/sh
 
 # 默认配置文件
-if [ ! -f "/app/conf/rtorrent.rc" ];then
-    cp /opt/conf/rtorrent.rc /app/conf/
-fi
-
-if [ ! -f "/app/conf/rutorrent.php" ];then
-    cp /opt/conf/rutorrent.php /app/conf/
-fi
+[ ! -f "/app/conf/rtorrent.rc" ] && cp /opt/conf/rtorrent.rc /app/conf/
+[ ! -f "/app/conf/rutorrent.php" ] && cp /opt/conf/rutorrent.php /app/conf/
+[ ! -f "/app/conf/rutorrent.conf" ] && cp /opt/conf/rutorrent.conf /app/conf/
 
 # 默认密码 123456
-if [ ! -f "/app/conf/httpPassword" ];then
-    echo "admin:zOfptPkebiKR." > /app/conf/httpPassword
-fi
+[ ! -f "/app/conf/httpPassword" ] && echo "admin:zOfptPkebiKR." > /app/conf/httpPassword
 
 # 每次运行都是用用户配置覆盖
 cp -f /app/conf/rtorrent.rc /root/.rtorrent.rc
 cp -f /app/conf/rutorrent.php /app/ruTorrent/conf/config.php
 cp -f /app/conf/httpPassword /etc/nginx/httpPassword
+cp -f /opt/conf/rutorrent.conf /etc/nginx/sites-available/rutorrent.conf
 
 # 权限修正
-chmod 644 /etc/nginx/httpPassword /root/.rtorrent.rc /app/ruTorrent/conf/config.php
+chmod 644 /etc/nginx/httpPassword \
+		/root/.rtorrent.rc \
+		/app/ruTorrent/conf/config.php \
+		/etc/nginx/sites-available/rutorrent.conf
+
 chmod -R 777 /app/*
 
 # 解锁
